requirements.txt
asgiref==3.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.5
django-filter==2.4.0
djangorestframework==3.12.2
drf-yasg==1.21.6
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
pytz==2023.3
PyYAML==6.0
requests==2.31.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.0.3
